+++ FribergWebAPI/Controllers/AgenciesController.cs
@@ -5,8 +5,8 @@ using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
-using FribergWebAPI.Data;
 using FribergWebAPI.Models;
+using FribergWebAPI.Data.Interfaces;
 
 namespace FribergWebAPI.Controllers
 {
