--- FribergWebAPI/Data/IResidence.cs
@@ -1,14 +1,0 @@
-using FribergWebAPI.Models;
-
-//author: Christian
-namespace FribergWebAPI.Data
-{
-	public interface IResidence 
-	{
-		Task<IEnumerable<Residence>> GetAll();
-		Task<Residence> GetById(int id);
-		Task Add(Residence residence);
-		Task Update(Residence residence);
-		Task Delete(int id);
-	}
-}
