+++ FribergWebAPI/Data/Interfaces/IAgency.cs
@@ -1,6 +1,6 @@
 ﻿using FribergWebAPI.Models;
 
-namespace FribergWebAPI.Data
+namespace FribergWebAPI.Data.Interfaces
 {
     public interface IAgency //author: Johan Krångh
     {
