+++ FribergWebAPI/Data/Repositories/CategoryRepository.cs
@@ -1,7 +1,8 @@
+using FribergWebAPI.Data.Interfaces;
 using FribergWebAPI.Models;
 using Microsoft.EntityFrameworkCore;
 
-namespace FribergWebAPI.Data
+namespace FribergWebAPI.Data.Repositories
 {
     public class CategoryRepository : ICategory
     {
